Internal map keys
The map data structure and associated functions are updated to distinguish
between the supplied form of keys and the normalized form.  The latter are
made internal to each map -- neither shared (e.g. with a loop object) nor
exposed via API functions.  For the time being, the normalized form is
retained in the data structure, but if it proves to not otherwise be used
then it may ultimately be removed.