Behavior change for cif_container_get_value()
Modified the documentation and implementation of cif_container_get_value()
to remove the behavior of invoking cif_value_clean() on the value pointer
referenced by its second argument under certain circumstances.  The function
has no good way to determine whether it is actually safe or appropriate to
do that, so one way or another we have to rely on the caller anyway.  With
the revision, the caller can cause a memory leak, but not likely a segfault.