Tests and corrections for cif_loop_get_names()
Added tests an a minor correction for cif_loop_get_names().  This function
is also tested, slightly more strenuously, in the course of testing various
other functions.  This test added here focuses on failure behavior, and in
that vein, the function under test was given the ability to diagnose invalid
loop handles.