Flagged intentionally-unused parameters in ciffile.c
The new CIF formatter is built on top of the more general (and also new) CIF
walker.  Because the walker is more general, its interface requires callback
function parameters that in some cases are not needed or used by the formatter.
These parameters are now flagged for documentary purposes, and to suppress
compiler warnings from those compiler for which we support doing so.