Changed the parse options to reflect CIF 2.0 going UTF-8-only
Removed the "suppress_utf8_fallback" parse option as CIF 2.0 is now
specified as UTF-8 only.  Added a "force_default_charset" parse option
that can be used in conjunction with "default_charset_name" to direct
the parser to use a specific charset without regard to any
counterindications, such as encoding signatures or system defaults,
including for CIF 2.