Adapted ICU unicode converter callbacks to the parser error callback API.
The built-in parser implementation relies on the ICU Unicode converter API,
which has its own callback mechanism for reporting encoding errors.  This
update introduces an ICU converter callback that wraps a CIF API callback
to signal encoding errors, along with new CIF error codes with which to
signal encoding errors.  Some rearrangement of the parser code and data
structures was required to make it work.