Provisional completion of parser callback support.
All callback calls are in place and additional support functions are created
where appropriate, especially with respect to handling cif_loop_t objects
for loops that have not yet been added to their containers.

This code has not yet been built and probably breaks the build.  It is being
committed to memorialize its current state.