Fixed buggy use of UCHAR_DECL / UCHAR_INIT
Use of the size argument to to the UCHAR_DECL() and UCHAR_INIT() macros
was inconsistent and in many places incorrect.  If NUL-termination of
the Unicode strings so declared is desired, then the given size must
account for it, but that was not consistently done.  In many cases, the
code worked anyway on the test platform, however, thus masking the
problem.  This commit updates all the tests to ensure that the static
Unicode strings they declare will be NUL terminated.