Reduced UTHash multiple evaluation of macro parameters.
Modified uthash.h to eliminate multiple evaluation of some hash parameters.
Of particular import here is the key length argument to HASH_ADD_KEYPTR, which
was evaluated many times and which is typically passed a call to the U_BYTES()
macro that counts the number of bytes in a Unicode string.