Added 'const' qualifiers to map key output variables.
Added a 'const' qualifier to the output variables in which table keys and
packet data names are returned, to emphasize that the returned keys must not
be modified, and to help the compiler identify violations.  Modified library
and test code to avoid triggering warnings associated with the change.
Cleaned up test code to avoid several additional warnings as well.