Added additional tests for cif_container_set_value()
Added tests and documentation for behavior when the value set via
cif_container_set_value() is NULL.  Made behavior consistent in that case
regardless of whether the item is intially present in the container.  Also
added a test for behavior when the item name specified to
cif_container_set_value() is not a valid CIF data name.