Minor build system tweaks.
Updated the Automake files to ensure that the new value assertion header
among the test sources is included in the distribution package.

Also, modified configure.ac to put macro quotes around the delimiters in
the bug report address, to make that address somewhat resistant to being
scraped from the file.  Because it is done via macro quotes, the
generated configure script is not affected.