Modified free function return types and add function attributes
Modified the cif_*_free() functions to return void instead of int, as the
underlying free() function does itself.  Also added GCC function attributes
to assentially all functions, and made it easier to perform similar work
in the future by wrapping function declarations in a macro.  The Doxygen
configuration needed to be updated to cope with the latter.