Fixed memory leaks in the map code and the packet iterator tests.
The packet iterator tests simply forgot to free a resource, but the map code
required more extensive changes.  Basically, the map entry metadata (keys)
could be leaked when an entry was removed from the map.  This certainly
affected operations on packets, and it probably also affected operations on
table values.