Added tests for modifying loops via an iterator.
Added tests focusing on ordinary use cases of cif_pktitr_update_packet() and
cif_pktitr_remove_packet().  Fixed a couple of small but impactful bugs in the
iterator code to make it pass the test.  Fixes required changes to other
code that records value objects in the DB.  Removed a duplicate SQL statement
from the library's repertoire, and changed the labels by which some others
are identified.