Minor SQL tweaks
Modified the query for updating values in packets to handle both the case
of an existing value being present and the case of an implicit unknown value
with the same query.  This allowed a slight simplification of the packet
update logic in the packet iterator code.

Also tweaked the schema slightly.  The only semantic difference was to
remove cascading updates for data name updates.  No other CIF identifiers
support cascading updates, and there is anyway no API-level support for
changing the name of an existing data item.