Added tests for cif_container_prune()
Added unit tests for cif_container_prune(), and corrected invalid SQL
that the tests revealed it trying to use.

The test process also revealed that the Valgrind suppressions file in
the distribution was only appropriate for platforms with 32-bit data
pointers; a new version is provided for platforms whose data pointers
are 64 bits wide.