Added tests for parsing fairly simple loops.
Added tests for parsing simple loops into an empty in-memory CIF object.
The parser itself worked flawlessly, but the tests did reveal a bug in
function cif_container_get_all_loops() in which it left its DB
transaction open on success.

Also took the opportunity to clean up and amend the documentation of
packet iterator lifecycle.