Code cleanup in parser.c
Renamed macro PUSHBACK_CHAR() to BACK_UP() to better distinguish it from
macro PUSH_BACK().  The two cannot easily be combined, as they do
different things: the former backs up the scanner itself so that it will
re-scan the most recently scanned code unit, whereas the latter causes a
specified number of already scanned and consumed code units to be taken
as the first few of the next token (without re-scanning them).