Made the parser flag errors for CIF 1 documents beginning with a BOM
The previous work on fixing BOM handling relaxed error detection a bit
too much.  CIF 1 documents should never contain the BOM character, nor
any of the bytes in its UTF-8 encoding.  The parser was nevertheless
accepting an initial BOM without complaint regardless of CIF version
(which can be unknown when the BOM is first encountered).  This update
causes the parser to emit an error -- which a handler can safely ignore
-- if a CIF 1 document starts with a BOM.