Added support for nested save frames
Made the relatively small changes needed to support nested save frames
in the data model and parser.  No specific tests for use of this feature
are yet in place, but the core dictionary file uses nested save frames,
so that test, at least, exercises the new capability to some degree.
Unfortunately, the core dictionary test is currently failing on account
of use of triple-quoted strings by the file.