Prevented nested frames in CIF 1 mode
Added back some code to the parser to flag an error for
apparently-nested save frames when parsing CIF 1.  The recovery path
assumes a frame terminator whenever one is missing before a frame header
frame header, thereby forcing all frames to be direct children of the
host data block (but only in CIF 1 mode).