Updated cif_walk() to support nested save frames
The previous modifications to add support for nested save frames
neglected to update the cif_walk() function.  This update makes
cif_walk() aware of nested frames.  The result is in fact a bit of a
simplification, as the internal function for traversing save frames was
coalesced with the one for traversing data blocks.