Multiple warning and compatibility cleanups
Performed widespread small changes to clear up compiler warnings and to
ensure that it builds and works in C++ (via g++) as well as in C89 (via
gcc).  In some cases (generally revolving around switch statements) the
code was documented instead of changed to avoid certain warnings.

Also fixed an assue in the 'configure' program that interfered with
configuring to use g++ as the C compiler.