More thorough fixup for ssize_t and ICU pkg-config
Ensured that all source files that reference type ssize_t both include
the config.h header (supposing that it is being used), and to include
unistd.h if it is available.

Switched the order in which ICU pkg-config packages are tested, and
suppressed error messages emitted by pkg-config when a non-existent ICU
package is requested.  (The configure script still outputs a diagnostic.)