Sped up adding packets to large loops
Implemented a revised mechanism for assigning loop row (packet) numbers,
and adjusted some triggers to reduce the work needed to add a packet to
a loop that already contains many of them.  Overall parsing time for one
test case was reduced roughly by a factor of 30!  The trade-off is that
row number management is now a little less automatic, and constraints
around the number of rows in a scalar loop are now enforced by the
C code instead of by triggers in the DB.