Upgraded autotools
Upgraded the versions of the autotools with which the build system is
built; this changes both the included autotools components and the
distributed components built from them.

Modified the configure flags used during a distcheck: the new Automake
doesn't like that when documentation is rebuilt, 'make distclean' does
not subsequently remove the generated docs.