Added Windows compatibility provisions
Added provisions for building on Windows.  The implementation was
developed and tested with the MinGW-w64 toolchain, but some changes are
more general.

SPECIAL NOTE: Windows' implementation of the STANDARD LIBRARY function
tmpfile() is brain-dead: it always attempts to create the temp file in
the root directory of the system drive, which should not be writable by
ordinary users.  A Windows-specific workaround is now conditionally
compiled in.