Added cif_get_api_version()
Added function cif_get_api_version(), and reconfigured the build system
to maintain the version code automatically based on the one in
configure.ac.  Tweaked the build system to include the new version.h
header and the old schema.h header in distributions, and to clean them
only during a 'maintainer-clean', yet to build them as needed.