Added the HTML documentation to the repo
Added the HTML documentation to the git repository.  It was always
included in the source tarball, but had previously been omitted from the
repo on account of being generated files.  Now that the project has been
put on GitHub, it is advantageous for the docs to be directly accessible
in the repository.