Added support for CIF 1.1 output format
Added support for output in CIF 1.1 format, plus a test case for that
behavior.  There is no new function; instead, the write options
structure is given a member that signals the output version requested.

The implementation hooks the existing writer framework to output a v1.1
version header instead of a v2.0 header, to suppress generation of
triple-quoted strings, and to reject Lists, Tables, and character
values containing text-field delimiters.