Fixed Bug #3: "_" being accepted as a data name
Fixed bug #3 in which the string consisting of a single underscore is
accepted as a valid data name by those functions that validate data
names, including at least:
  cif_container_create_loop()
  cif_container_set_value()
  cif_loop_add_item()
  cif_packet_create()
  cif_packet_set_item()

Tests were added for this behavior for the above functions.  A one-line
fix for all of the above was applied to internal function
cif_validate_name() in utils.c.